Update palettes & highlights
The draft WCAG3 / APCA color contrast calculations includes better
consideration of readability, especially for light-on-dark. This was
used to touch up the theme colors, especially for dark backgrounds.

In addition, the darker half of the base tone palette was skewed a
little bluer.

Style tweaks include:
- Replaced `VertSplit` with thinner `WinSeparator`
- Consistent background for floats

This tool was very useful for the color contrast calucations:
    https://cliambrown.com/contrast/